requirements.txt
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
h11==0.16.0
idna==3.10
import-ipynb==0.2
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.1
referencing==0.36.2
rpds-py==0.26.0
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.4
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
